Promote from_parts to a validated public constructor; make raw kernels pub(crate) (#141, #142)
Establishes a single validated boundary for constructing an index from
already-decoded parts and removes the unchecked low-level kernels from the
public surface.

`TurboQuantIndex::from_parts` becomes `pub` and returns
`Result<Self, FromPartsError>`, checking every structural invariant at one
chokepoint: bit_width in {2,3,4}; a committed dim a positive multiple of 8
and <= MAX_DIM; packed_codes / scales / TQ+ array lengths; and the
lazy-state constraints. The paired accessors packed_codes(), scales(),
tqplus_shift(), tqplus_scale() are promoted to pub so an index round-trips
through external storage (the pg_turbovec / #70 use case). New public error
type FromPartsError; TurboQuantIndex now derives Debug.

The raw kernels codebook::codebook, encode::encode, pack::repack and
search::search become pub(crate). On the public surface they were a
soundness/DoS hazard: search::search read out of bounds / SIGBUS'd from
inconsistent caller lengths (UB from safe code, #141); encode/repack
panicked opaquely and codebook hung on an unbounded 2^bits allocation or
returned silently-wrong output (#142). Internal callers (the high-level
index types) keep calling them directly — they already establish the
invariants, so hot paths gain no per-call validation.

Breaking for the Rust crate (removes public items). CHANGELOG notes the
migration: build via from_parts or the high-level types.

Kernel-level correctness tests that reached the now-internal leaves moved
in-crate (src/kernel_tests.rs); the dump_state dev example, which existed
only to dump the internal codebook, was removed. New tests/from_parts.rs
covers the happy-path round trip (byte-identical persistence) and every
invariant rejection, including bits=40 rejected cheaply without the 2^40
allocation.
